Drop use of `something` in reflection.jl
reflection.jl is included in Core.Compiler, but the `something` function is not, so we cannot use it here.
Throw an explicit error instead. I considered having it not throw the error when `err==false`, but it
seems to me like passing such a UnionAll is just a user error, rather than an acceptable query, so
I left it as is and simply did the bug fix.