Help for professional programmers
I just started learning Julia a month ago and the getting started documentation is really just an overview of some basic elements.  It needs to answer the question: how do I structure my code for using projects.  The JuliaCon 2016 talk on programmers vs scientists touches on this.  I agree that Julia's tools are extensive.  But the getting started docs needs to offer initial guidance for programmers on how to develop projects right away.   Even just a "here's how I do it" would be extremely helpful.  The reference I gave is almost too short.  But even that is very valuable.