Add typedef jl_thread_t, fix jl_wakeup_thread
POSIX does not mandate that pthread_t is an integer, and in fact on FreeBSD it
is a pointer. While casting a pointer to an unsigned long and back works there
right now, it seems safer on the long run use the correct type.

Also fix a bug in jl_wakeup_thread where the native thread id of the owner
of jl_uv_mutex was truncated to 16bits.