rand(BigFloat): create BigFloat with precision according to the sampler
If the global precision for BigFloat changed between creating the sampler
and generating a random BigFloat, bad things would happen, in particular
if the precision was decreasing (we would write into arrays out-of-bounds).
So create a fresh BigFloat without consideration for the current
precision, only for the precision stored in the sampler.