Solving Lyapunov equation for matrices given in Adjoint form
There seems to be no consensus about what should be the Lyapunov equation: A'P+PA+Q=0 or AP+PA'+Q=0. For instance, Wikipedia says differently than the doc of `lyap`. For this reason, I think it would be nice to have an easy interface to switch between both definitions by allowing to put `A'` directly in the arguments of `lyap`.