[loading]: Fix new tests post-Preferences merge
The new Preferences loading strategy changed some pieces of the loading
machinery, in particular, it added another parameter to
`Base.compilecache_path()`.  This was not caught by CI because the
Preferences PR stood open for so long that the other tests changed
after the CI had already turned green, and we don't re-run tests when
another branch is merged, so we often have stale tests.