Profile: fix timeout units
Previously, we were using units of seconds (rounded down) but the
function expected milliseconds. So this was effectively always zero
(ignored) in the computation.

Empirically, the documentation for Sleep is fundamentally incorrect, and
the stated advice to add `wPeriodMin` is bad. So here we remove that.

Also update the Profile code defaults to sample for longer, and faster,
since only the Win32 stack-unwinder (dbghelp) is absurdly slow.