A faster version of exp, exp2, exp10 for Floating point numbers (#36761)
* A faster version of exp for Float64

This is based on the Glibc algorithm which at-chriselrod (Elrond on discourse) described the algorithm of for me. It appears to be about 2x faster than the current algorithm, and equally accurate over the range for which I have tried it. It also theoretically should be easier to vectorize as branches are only used for checking for over/underflow.