Fix lock leak in iosteam test
The `resize!` call can throw if the array is not resizable (we even
have a test for it). However, this would leak the lock, since the
`_lock_ios` macro does not have an implicit try/catch, since it's
really only supposed to lock the C calls. Fix that by adding the
necessary try/catch around the resize call. Perhaps in the future
the compiler will be able to automatically elide try/catch blocks
around calls known not to throw, in which case, we could just put
it in the macro, without inviting performance problems.

Fixes #37218