win,Profile: fix timeout units (#38208)
Previously, we were using units of seconds (rounded down) but the
function expected milliseconds. So this was effectively always zero
(ignored) in the computation.

Empirically, the documentation for Sleep is fundamentally incorrect, and
the stated advice to add `wPeriodMin` is bad. So here we remove that.
(Microsoft has given the thumbs up to this observation also now.)

Also update the Profile code defaults to sample for longer, and faster,
since only the Win32 stack-unwinder (dbghelp) is absurdly slow.

Co-authored-by: kimikage <kimikage.ceo@gmail.com>