[AbstractInterpreter] Implement callbacks for cache invalidations
AbstractInterpreter allows down-stream compiler variants to implement
their own caching of inference results. This leads to the issue of that
invalidations of methods are not propagated to those caches, leaving
downstream implementations with their own variant of 265, with limited
ability to mitigate it.

This PR extends the invalidation scheme by allowing downstream
implementation to attach invalidation callbacks to MethodInstances.
MethodInstances are used as the key to cache and can therefore be used
to walk and invalidate the user cache.