Hack in acquire-load of task state to fix tests (#38557)
On x86 loads have implicit acquire semantics, but the same is
not true on other architectures (e.g. aarch64). We've been
observing test failures on aarch64 as a result (#38555).
Currently, we don't really have a good way to modify fields
in an atomic way (pending #37847), so this just achieves
the same thing using llvmcall to try and get the tests passing.