Add pretty-printing for Core.Compiler.Timings objects
* Add minimal printing for Core.Compiler.Timings.Timing

These objects are huge when printed with `show_default`.
This makes the printing very brief. More detail is available with
with AbstractTrees.print_tree.

* Add nice printing for Core.Compiler.Timings.InferenceFrameInfo

This inlines the `Core.Const` args into the signature, which
results in a very compact way to see both the type and the value, when applicable.

Co-authored-by: Nathan Daly <NHDaly@gmail.com>