inline exp functions
Now that the bulk of my changes to exp have been merged, I want to know if we should consider adding `@inline` to the code. The upside is faster performance and possible vectorization (although vectorizing requires `@simd ivdep` for `Float64`). The disadvantage is probable increased compile times, and possible performance hits in some situations due to spilling code caches. This `@inline` was removed from the original PR as it added questions to a PR that was already hard to merge due to it's scope, but I think this is probably worth it.