Iterator interface: Change i to item in example
Many iterators have an internal counter of iterations for the abort condition, e.g. `Squares` in the example. When implementing such an iterator while following this example, the naming can be a bit misleading because `i` in the example is _not_ a counter but instead the next item of the iteration. However, `i` is also commonly used as a counter in iterations.

I think renaming it to `item` makes it a bit clearer.

(cc @miguelraz)