[libjulia] correct soname to use only major version number
The soname was changed in #38160 to be libjulia.so.1.6 instead of
libjulia.so.1, but probably not on purpose.

This contradicts what is written in the comment in Make.inc and it breaks
e.g. libcxxwrap_julia_jll which is linked against libjulia.so.1.
See https://github.com/JuliaPackaging/Yggdrasil/issues/2233