optimize invokelatest call performance
People previously were making dumb work-arounds like try/catch around
this to get a bit faster here, since the compiler couldn't optimize it
before. Applications typically shouldn't use this function in
performance sensitive places, as it hints that their design is flawed,
but might as well make it faster anyways.

and optimize invokelatest kwcall too, while we are at it