Execute `backtrace` once before testing formatting (#38886)
On some platforms (PowerPC) the call to the `jlplt` is not a tail-call
and so it will be part of the backtrace. This means we are off-by-one
and won't skip the Julia function `backtrace` messing up tests that
check precise formatting.

(cherry picked from commit fc577d0b41a108ce6c502d80271292c1e442b18e)