[cli/trampolines]: Fix `aarch64-apple-darwin` trampoline ASM syntax
It turns out that aarch64 assembly syntax uses `;` as the comment
character, not as the statement separator.  So we need to polyfill
that.  We also set an alignment of 4 bytes (`2^2`) which is required on
Apple targets, but is also a good idea on most other aarch64 machines.

(cherry picked from commit cdb0833424760f1369facf4ee9330d9f43486b65)