Assert that _artifact_str returns a String (#38975)
I have no idea whether this is important, but I noticed this while investigating triggers for inference in a package. Feel free to close if you think this is irrelevant.

Incidentally, `_artifact_str` is a ~60ms precompile (that's just the inference, not the codegen). I see that its precompile was commented out in 924484f2f00; just mentioning it in case there's an opportunity to squeeze out a bit more latency-reduction.