Refactor JL_GC_PUSHFRAME to internalize layout assumptions (#38982)
Right now callers of JL_GC_PUSHFRAME need to be aware that the
roots start three pointers after the end of the struct. That
seems like an implementation detail of the macro, so refactor the
macro to also take the assignment location for the locals instead.
In the process fix an error where the gc-analyzer version of this
code assumed a different layout.