Reference functions instead of methods
By referencing methods sometimes and not just functions, there was some duplication and confusion. For instance, we had `get(::Any, ::Any, ::Any)` method and also a `get` function, the latter of which duplicated the documentation for the former.

Is there an unseen benefit to referencing just the function at times?

If this PR is acceptable, I can go through the other pages to do the same.