Small drive-by cleanup: Make constant explicit
This code is a bit of a leftover from when we still had `nfields(::Type)`,
but at this point it was always just returning the number of fields
of `DataType`. I was a bit confused by this code, so I thought
just using `jl_datatype_t` here explicitly makes clear that it does
actually just return said constant.