Improve copyto! for short heterogeneous tuples (#39035)
There was a complaint on Twitter [1], that Julia is slower than
Ruby(!) at allocating short heterogeneous arrays. The code looks
something like this:

```
f() = for i in 1:100000; b=[1,2.0]; end;
```

Of course, this benchmark is silly because the arrays are
unused, so the entire code should be dropped once we have
further improvements to our lifetime analysis, which will
change the runtime 0. However, since we're not quite able
to do that yet, it is somewhat unexpected that we should
be slower than Ruby here. Admittedly our arrays are tuned
for larger sizes, but still. Upon further investigation,
it turns out that the issue is that we're allocating extra
boxed objects when pulling the fields out of the literal,
since we don't have any specialization for `copyto!` from
tuples, falling back to the generic `copyto!` method, which
is type unstable in the presence of heterogeneous tuples.

Since this code is reachable from surface syntax, I think
it's worth putting in a couple of specializations to
accelerate the array construction, which is what this PR
does (up to tuples of length 10 to avoid excessive
codegen). Since the type-instability issue applies whether
the value is used or not, this should still be useful even
if we're able to prove that the allocations can be dropped
entirely.

Before:
```
julia> @benchmark f()
BenchmarkTools.Trial:
  memory estimate:  21.36 MiB
  allocs estimate:  600000
  --------------
  minimum time:     14.928 ms (0.00% GC)
  median time:      16.297 ms (0.00% GC)
  mean time:        16.323 ms (3.84% GC)
  maximum time:     26.042 ms (0.00% GC)
  --------------
  samples:          307
  evals/sample:     1
```

After:
```
julia> @benchmark f()
BenchmarkTools.Trial:
  memory estimate:  9.16 MiB
  allocs estimate:  100000
  --------------
  minimum time:     2.117 ms (0.00% GC)
  median time:      2.276 ms (0.00% GC)
  mean time:        2.455 ms (7.62% GC)
  maximum time:     4.638 ms (28.99% GC)
  --------------
  samples:          2036
  evals/sample:     1
```

Which puts us about on par with Ruby on my machine (not entirely
unsurprising, since this is essentially an allocation benchmark
at this point and both allocators are written in C).