Fix `p7zip_jll` environment setup
On most platforms, setting `LD_LIBRARY_PATH` and friends was not
necessary for invoking `7z` because we fell back to the natural library
search paths for things like `libstdc++`, but in truly minimal
environments (like Alpine linux) this does not work; so let's set these
paths properly.  We do this work manually in v1.6, in v1.7 we hope to
have a JLLWrappers-based solution.