inference: make Limited tracking part of the type lattice (#39116)
This helps refine our knowledge of the `[limited]` flag setting, which
previously would always exclude a result from the cache when hitting a
cycle. However, we really only need to exclude a result if the result
might be dependent on that flag setting. That makes this formally part
of the lattice, though can be annoying to work with yet another wrapper,
so we try to add/remove it late/early to propagate it when necessary.

(cherry picked from commit 5f10eb9085fd5a622430d020c8bc9141c7acb7c7)