Fix inference of `Argument`
I was pretty sure I had made this commit already, but I guess
it never made it onto master. Regardless, in d6d5208, we switched
Argument to be a legal value in CodeInfo and gave it semantics
equivalent to `Slot`. This is useful to avoid back-and-forth
conversion. However, we missed one place in inference which
assumed that literal arguments would just end up as Consts,
which is of course wrong - causing crashes. Now usually we
don't end up with literal `Argument`s in CodeInfos, since
the frontend doesn't produce them, but they can come
in when trying to generate SSA code in an external
package and then converting back to `CodeInfo` (which
after the above commit no longer turns the `Argument`s
into Slots).