Fix ~700 PyPlot invalidations (#39420)
PyCall specializes a lot of low-level methods like
`convert(::Type{Symbol}, arg)`, and this triggers a lot of invalidations.
The worst are for keyword-argument functions, but there are quite
a few others. This drops the number from ~1100 to ~400.
It's quite likely that additional improvements could be made, but this
is major progress.

Related: #39419.