Run jl_resolve_globals_in_ir on the result of generated functions (#39483)
We didn't use to run this when the generated function returned a CodeInfo (though we did run it for AST). However, now that this also does the conversion of :opaque_closure_method to actual Method objects, I think the best thing to do would
be to run it here to avoid the generated function having to construct its own Method object from scratch.