Disable Profiler stack switch heuristic when running under rr
rr uses a separately allocated stack to avoid disturbing the main application.
It has a carefully crafted set of CFI information to make sure that unwinders
can successfully unwind through this stack switching. However, our unwinder
has a bail out path when this stack switching is detected. This is sensible in
general (assuming we're dealing with unknown external code with potentially
bad unwind info). However, when running under rr it can cause failures in the
Profile test, so just disable it there.