test: fix small race in AsyncCondition test
There was a small race window where the second t might not have finished
exiting, leading to interference in the Workqueue, resulting in
process_events from running again before the async callback had a chance
to run and dequeue the first event. (This test is probably much stronger
than necessary.)

N.B. This depends on the recent changes in
af353ecb76310589f3baf3e746ee35c7e894672d for being able to extract the
internal state needed for the sake of the test.

Fixes #39527