[loading] Search in Sys.STDLIB if name/uuid is known. (#39572)
Locating a package with known uuid means the uuid was either
found in a manifest in LOAD_PATH or a "package folder" in LOAD_PATH.
However, looking up the package entrypoint fails for stdlibs that
are not in the sysimage, even if they have been explicitly added
to project/manifest by the user. This patch allows looking in
Sys.STDLIB after stdlibs, fixes #39504.