Add `JULIA_PKG_SERVER` to Environment Variables in docs
I'm not really sure what to write in here. I'm not super familiar with this ENV variable, nor do I know how best to explain it.

I'd love some help with this!! Please feel free to edit this PR directly!!

I am opening this because we recently needed to disable the package server (we were having consistency issues, where the registry wasn't updating), and we couldn't find documented anywhere what the Environment Variable is to do that. We eventually found it in the julia slack. I'd like to add it here so others can find it. :) thanks!