improve constraint propagation with multiple ||
What this PR does is expand for example
```julia
if x === nothing || x < 0
    return 0
end
```
to
```julia
if x === nothing
    @goto l
else
    if x < 0
        @label l
        0
    end
end
```
in lowering, which IIUC should correspond to the first option @vtjnash proposed in https://github.com/JuliaLang/julia/pull/39549/files#r573995564. Are there any potential problems with emitting `goto`s here? Or is there a nicer way to fix this than in `expand-if`?
fixes #39611