Switch macOS task mechanism back to ASM (#39679)
5327824ec8c452410e2a2f755921764df9344855 rearranged the ifdefs here and
switched macOS to unwind based switching, which is somewhat slower. That was
deemed acceptable at the time, because it work around some GDB misbehavior
in setjmp. However, we're about to update to the LLVM libunwind fork, which
appears to be even slower here, so switch back to ASM to avoid that performance
regression.