RFC: expose expand-forms
I think this is not only useful for more fine-grained introspection, but I also see quite a few macros reimplementing a lot of this logic for certain transformation. One such recent example is #37379, where we might just want to remove a lot of the special cases for `@which` and similar macros and instead expand the expressions first.

One concern of course would be how much we want to expose implementation details of the frontend, as this might complicate future overhauls, but I would assume that even if we wanted to rewrite lowering in Julia at some point, we would likely still have some kind of desugaring pass over the AST, so I don't think this constrains us much in that regard.

I am not sure the module argument is actually needed. IIUC, this really only serves to resolve `(top ...)` and `(core ...)` to `Base` and `Core` accordingly, so we can maybe just always pass `Base` here.