Add support to multiple property pairs in sprint (#39381)
* Add support to multiple property pairs in sprint

Sometimes it is required to pass multiple properties to `IOContext` in
`sprint`. For example, if we want to print with `:compact` and `:limit`
set to true. Currently, the only possible way to do this is creating an
`IOContext` using a dummy `IOBuffer` with those parameters. Hence, this
commit allows to pass a vector of pairs `:key=>value` to `context`
keyword of `sprint` so that we can easily set multiple properties.

This is not a breaking change, and no performance regression was
identified when using the previous function signatures.

* Add compat annotation to sprint

* Update base/strings/io.jl

Co-authored-by: Rafael Fourquet <fourquet.rafael@gmail.com>

* Update sprint docstring

* Update NEWS.md

* Move NEWS.md entry to the correct place

Co-authored-by: Rafael Fourquet <fourquet.rafael@gmail.com>