Allow inlining of code for big pure results
In some cases we can statically compute the result of a
pure call, but we refuse to inline it into the IR because
it would be too big. In these cases, we'd still like to be
able to do regular inlining, so augment MethodResult pure
to forward the call info for such calls rather than relying
on inlining to recompute the method match.