Make ReadOnlyError on PROT_NONE consistent (#39803)
This replaces #36625. Essentially the question is what should
reading from PROT_NONE memory do. At the moment we throw
ReadOnlyMemoryError on Linux and pass through the segfault on
Windows/Mac. It's not super clear what the right answer is,
though it would be nice to be consistent. After some discussion,
it seemed that not throwing ReadOnlyMemoryError on Linux was probably
the better consistency direction for two reasons:

1. Getting a ReadOnlyMemoryError is confusing if the failing operation
   is a read. We could of course rename the error, but that would be
   a bigger breaking change.
2. The purpose of this exception is to protect people who accidentally
   try to write to mmap'ed memory that they don't have write permissions
   to. It was never intended to become a general memory access error
   exception, so it seems prudent to restrict it as much as possible.

To make this happen, there is some platform specific code here to
read the memory error register out of the signal context.
Decoding this register is shared code between the operating systems,
but the register itself is in an OS-specifc location in the signal
frame. I have implemented the retrieval code for all our CI'ed
platforms. Other users will get an appropriate `#warning` (similar
to other code in this file) and can fill in the necessary code
at their leisure. A new test in test/misc.jl will indicate whether
or not the check is correct.