llvm-alloc-opt: handle dead code better (#39801)
In some cases (particularly after removing a phi node), we might end up
in a circumstance where it appears statically that we would use a slot
for both a ref and bits. Avoid generating malformed IR in this case.

In the future, this situation could also possibly happen if we walked
through phi nodes and attempted to merge the contents (with great care).