[Filesystem]: `rm(;force=true)` should ignore `readdir()` errors
In my ever-expanding quest to test the limits of filesystem permissions,
I have found that if you create subdirectories without read permissions
inside of your temporary directory, the `mktempdir()` cleanup routine
will fail due to `rm()` attempting to `readdir()`, despite the fact that
it can remove an empty directory without reading it just fine.  This is
in fact what `rm -rf` does, and so here we enable this behavior for
`rm(;force=true)`.  I discovered this beahvior when attempting to mount
an `overlayfs` work directory within a temporary directory.