make broadcasting into getproperty extensible (#39473)
* make broadcasting into getproperty extensible

Don't know if `dotgetproperty` is a good name for this, since as discussed in #36741, the way this works is a bit different from `dotview`. Right now, `dotproperty` returns a function which is used instead of `getproperty`, but not sure if that is a bit strange and we should just make it behave like `dotview` instead.

fixes #36741

* make dotgetproperty call getproperty directly

Co-authored-by: Jameson Nash <vtjnash@gmail.com>